feat: open notifications in right-side modal on desktop
On desktop the header Alerts button now opens the full notifications list in
a right-loading modal (matching the My LearnCard profile modal pattern) instead
of navigating to the /notifications page. Mobile keeps the existing full-page
route.

- Extract shared NotificationsListView from NotificationsPage so the page and
  modal render identical list content
- Add NotificationsModal (right-modal presentation with tabs + close button)
- Add useOpenNotifications hook: desktop opens the modal, mobile navigates to
  /notifications
- Add optional onOpen prop to NotificationButton so the app layer can drive the
  modal without base depending on app code
